The logic that loads neighbor pages when empty space occurs in visible regions is changed
and it is separated into its own function, so that it can be called when required.
Previously, the logic was inside mouse_wheel_in_canvas event handler.

Now, the function is called in the above event handler, and also,
when annotations are navigated using down and up keys

On branch main
Changes to be committed:
 modified:   main.py